Revert "pages added"
