+++ package.json
@@ -6,7 +6,6 @@
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.15.9",
-    "@mui/lab": "^5.0.0-alpha.165",
     "@mui/material": "^5.15.9",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
@@ -21,7 +20,7 @@
     "yup": "^1.3.3"
   },
   "scripts": {
-    "start": "set PORT=3030 && react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
