+++ src/librarys/MUIIcons.js
@@ -2,6 +2,5 @@ import { PhotoCamera } from '@mui/icons-material';
 import CloseIcon from '@mui/icons-material/Close';
 import PersonAddIcon from '@mui/icons-material/PersonAdd';
 import AddIcon from '@mui/icons-material/Add';
-import AccessAlarmIcon from '@mui/icons-material/AccessAlarm';
 
-export { PhotoCamera, CloseIcon,PersonAddIcon, AddIcon,AccessAlarmIcon }
+export { PhotoCamera, CloseIcon,PersonAddIcon, AddIcon }
