+++ src/services/API_ENDPOINTS.js
@@ -8,17 +8,12 @@ const PROFILE = "profile";
 
 
 
-const SEARCH_USER = "search-users";
-const USER_LIST_FOR_CONNECTIONS = "user-list-for-connections";
-const MY_RECEIVED_CONNECTIONS = "my-received-connections";
-const MY_SEND_CONNECTIONS = "my-send-connections";
+const SEARCH_USER = "search-users"
+const USER_LIST_FOR_CONNECTIONS = "user-list-for-connections"
 
-// connections
-const SEND_REQUEST = "send-request";
 
 // Exporting all variables at once
 export {
     AUTH_LOGIN, AUTH_SIGNUP, PROFILE, SEARCH_USER,
-    USER_LIST_FOR_CONNECTIONS, SEND_REQUEST, MY_RECEIVED_CONNECTIONS,
-    MY_SEND_CONNECTIONS
+    USER_LIST_FOR_CONNECTIONS
 };
