--- src/user/ManagerInvitation/index.js
@@ -1,48 +1,0 @@
-import { useNavigate } from "react-router-dom";
-import {
-    Grid, Card, Typography, CardHeader, Button, Box,
-    Tab
-} from "../../librarys/MUILibrary";
-import { TabContext, TabList, TabPanel, } from '@mui/lab';
-
-import SettingsIcon from '@mui/icons-material/Settings';
-import { useState } from "react";
-import InvitationReceived from "./components/InvitationReceived";
-import InvitationSend from "./components/InvitationSend";
-const headerSX = {
-    '& .MuiCardHeader-action': { mr: 0 }
-};
-const ManagerInvitation = () => {
-    const navigate = useNavigate();
-    const [value, setValue] = useState('1');
-    return (
-        <Grid container spacing={3}>
-            <Grid item xs={12} sm={6} md={8}>
-                <Card>
-                    <CardHeader
-                        sx={headerSX}
-                        title={<Typography variant="h6" component="h6"> Manage invitations</Typography>}
-                        action={<Button variant="outlined" startIcon={<SettingsIcon />} onClick={() => navigate('/invitation-manager')}> Manage</Button>} // include the header action here
-                    />
-                </Card>
-            </Grid>
-            <Grid item xs={12} sm={6} md={8}>
-                <Card>
-                    <Box sx={{ width: '100%', typography: 'body1' }}>
-                        <TabContext value={value}>
-                            <Box sx={{ display: 'flex', justifyContent: 'space-between', borderBottom: 1, borderColor: 'divider' }}>
-                                <TabList onChange={(event, newValue) => setValue(newValue)} textColor="secondary" indicatorColor="secondary" aria-label="lab API tabs example">
-                                    <Tab label={"Received"} value="1" />
-                                    <Tab label={"Sent"} value="2" />
-                                </TabList>
-                            </Box>
-                            <TabPanel value="1"><InvitationReceived /></TabPanel>
-                            <TabPanel value="2"><InvitationSend /></TabPanel>
-                        </TabContext>
-                    </Box>
-                </Card>
-            </Grid>
-        </Grid>
-    )
-}
-export default ManagerInvitation;
