+++ src/user/MyNetwork/components/ConnectionInformation.js
@@ -9,8 +9,6 @@ import SendIcon from '@mui/icons-material/Send';
 import ExpandLess from '@mui/icons-material/ExpandLess';
 import ExpandMore from '@mui/icons-material/ExpandMore';
 import StarBorder from '@mui/icons-material/StarBorder';
-
-import { Link } from 'react-router-dom';
 const ConnectionInformation = () => {
   const [open, setOpen] = React.useState(true);
 
@@ -22,9 +20,7 @@ const ConnectionInformation = () => {
       <Card >
         <List>
           <CardHeader title="Manage my network" />
-          <ListItemButton
-            component={Link}
-            to={"/connections"}>
+          <ListItemButton>
             <ListItemIcon>
               <SendIcon />
             </ListItemIcon>
