+++ src/user/MyNetwork/index.js
@@ -1,30 +1,37 @@
-import React, { useCallback, useContext, useEffect, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
 
 import {
-    Grid, Stack, Skeleton, Card, Typography, CardHeader, Button
+    Grid, Card, CardContent, CardHeader, Avatar, Typography, Button, IconButton,
+    CardActions, Stack, Skeleton
 } from "../../librarys/MUILibrary";
+import { CloseIcon, PersonAddIcon, AddIcon } from "../../librarys/MUIIcons";
+import { styled } from '../../librarys/MUIStyles';
 
-import ConnectionInformation from '../../components/common/ConnectionInformation';
+import ConnectionInformation from './components/ConnectionInformation';
 import { getUserList } from '../../services/CommonService';
-import ConnectionCard from './components/ConnectionCard';
-import { sendRequest } from '../../services/ConnectionService';
-import { AuthContext } from '../../contexts/AuthContext';
-import SettingsIcon from '@mui/icons-material/Settings';
-import { useNavigate } from 'react-router-dom';
 
+const StyledAvatarContainer = styled('div')({
+    width: '100%',
+    textAlign: 'center',
+    margin: 'auto',
+});
+
+const StyledCardUp = styled('div')({
+    height: '70px',
+    backgroundImage: "url('https://ik.imagekit.io/hydlcbl5qlg/public/misc/28iK9OtJk9TlPTbYUyohzvplQminYXV8ZjfrLIiw.jpg')",
+    backgroundRepeat: 'no-repeat',
+    backgroundSize: 'cover',
+});
 
 const MyNetwork = () => {
-    const navigate = useNavigate();
-    const { user } = useContext(AuthContext);
-    const loginUserId = user.userInfo.userId;
+
     const [state, setState] = useState({
         users: [],
-        data: [],
         loading: true,
         page: 1,
         pageSize: 15,
     });
-    const { users, data, loading, page, pageSize } = state;
+    const { users, loading, page, pageSize } = state;
     const observer = useRef(null);
 
     useEffect(() => {
@@ -38,11 +45,11 @@ const MyNetwork = () => {
                 const { data } = await getUserList({ page, pageSize });
                 setState(prevState => ({
                     ...prevState,
-                    data: data.data,
                     users: [...prevState.users, ...data.data],
                     loading: false
                 }));
             } catch (error) {
+                console.log("error==>", error);
                 setState(prevState => ({
                     ...prevState,
                     loading: false
@@ -52,12 +59,11 @@ const MyNetwork = () => {
 
         userSearching();
     }, [page, pageSize]);
-
     const lastUserElementRef = useCallback((node) => {
         if (loading) return;
         if (observer.current) observer.current.disconnect();
         observer.current = new IntersectionObserver((entries) => {
-            if (entries[0].isIntersecting && data.length > 0) {
+            if (entries[0].isIntersecting) {
                 setState((prevState) => ({
                     ...prevState,
                     page: prevState.page + 1
@@ -65,54 +71,7 @@ const MyNetwork = () => {
             }
         });
         if (node) observer.current.observe(node);
-    }, [loading, data.length]);
-
-
-    const handleUserConnection = async (userId, index) => {
-        try {
-            const userInfo = {
-                _id: null,
-                requester: loginUserId,
-                recipient: userId,
-                status: "pending",
-                createdAt: new Date(),
-                updatedAt: new Date(),
-                __v: 0
-            }
-            const response = await sendRequest({ receiverId: userId });
-            if (response.status) {
-                setState(prevState => {
-                    // Clone the array of users
-                    const updatedUsers = [...prevState.users];
-
-                    // Clone the user object to update
-                    const updatedUser = { ...updatedUsers[index] };
-
-                    // Update friendRequestsReceived with the new request
-                    updatedUser.friendRequestsReceived = [
-                        ...updatedUser.friendRequestsReceived,
-                        userInfo
-                    ];
-
-                    // Update the user object in the array
-                    updatedUsers[index] = updatedUser;
-
-                    // Return updated state
-                    return {
-                        ...prevState,
-                        users: updatedUsers
-                    };
-                });
-            }
-        } catch (error) {
-            console.log("error==>", error);
-        }
-        console.log("handleUserConnection", userId);
-    }
-    // constant
-    const headerSX = {
-        '& .MuiCardHeader-action': { mr: 0 }
-    };
+    }, [loading]);
     return (
         <Grid container spacing={3}>
             <Grid item xs={12} sm={6} md={4}>
@@ -120,16 +79,39 @@ const MyNetwork = () => {
             </Grid>
             <Grid item xs={12} sm={6} md={8}>
                 <Grid container spacing={3}>
-                    <Grid item xs={12} sm={6} md={12}>
-                        <Card>
-                            <CardHeader
-                                sx={headerSX}
-                                title={<Typography variant="h6" component="h6"> Manage invitations</Typography>}
-                                action={<Button variant="outlined" startIcon={<SettingsIcon />} onClick={() => navigate('/invitation-manager')}> Manage</Button>} // include the header action here
-                            />
-                        </Card>
-                    </Grid>
-                    <ConnectionCard users={users} handleUserConnection={handleUserConnection} />
+                    {
+                        users.length > 0 && users.map((user, index) => (
+                            <Grid key={users._id} item xs={12} sm={6} md={4}>
+                                <Card>
+                                    <StyledAvatarContainer>
+                                        <StyledCardUp />
+                                        <CardHeader
+                                            action={
+                                                <IconButton sx={{ marginTop: '-120px', backgroundColor: '#623e09', color: 'white' }}>
+                                                    <CloseIcon color={"red"} />
+                                                </IconButton>
+                                            }
+                                        />
+                                        <Avatar sx={{ width: 150, height: 150, margin: 'auto', border: '2px solid #fff', marginTop: '-100px', zIndex: '11' }}>
+                                            <Avatar sx={{ height: '100%', width: '100%' }} src="https://avatars.githubusercontent.com/u/146355358?v=4" alt="woman avatar" />
+                                        </Avatar>
+                                        <CardContent>
+                                            <Typography variant="h5" component="h5" >
+                                                {user.username}
+                                            </Typography>
+                                            <Typography variant="body1">
+                                                {index}:- Lorem ipsum dolor sit amet,
+                                            </Typography>
+                                        </CardContent>
+                                        <CardActions>
+                                            <Button size="small" sx={{ width: '50%' }} variant="contained"><PersonAddIcon /> Connect</Button>
+                                            <Button size="small" sx={{ width: '50%' }} variant="outlined" ><AddIcon /> Connect</Button>
+                                        </CardActions>
+                                    </StyledAvatarContainer>
+                                </Card>
+                            </Grid>
+                        ))
+                    }
                     <Grid item xs={12} ref={lastUserElementRef}>
                         {
                             loading && (
