+++ src/user/layout/index.js
@@ -1,11 +1,10 @@
-import { Outlet, useNavigate } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
+import * as React from 'react';
 import { styled } from '@mui/material/styles';
 import { Box, CssBaseline } from '@mui/material';
 
 import { Header } from './Header';
 import { Sidebar } from './Sidebar';
-import { useContext, useEffect, useState } from 'react';
-import { AuthContext } from '../../contexts/AuthContext';
 
 const drawerWidth = 240;
 
@@ -40,10 +39,9 @@ const DrawerHeader = styled('div')(({ theme }) => ({
 }));
 
 function UserLayout() {
-    const navigate = useNavigate();
-    const { user } = useContext(AuthContext);
 
-    const [open, setOpen] = useState(false);
+
+    const [open, setOpen] = React.useState(false);
 
     const handleDrawerOpen = () => {
         setOpen(true);
@@ -52,12 +50,7 @@ function UserLayout() {
     const handleDrawerClose = () => {
         setOpen(false);
     };
-    useEffect(() => {
-        if (!user?.isLoggedIn) {
-            navigate('/auth/login')
-            console.log("user auto logout");
-        }
-    }, [user])
+
     return (
         <Box sx={{ display: 'flex' }}>
             <CssBaseline />
